Finished mapper

+++ remapper/src/main/kotlin/org/spectral/remapper/asm/ClassGroup.kt
@@ -11,38 +11,34 @@ import java.util.jar.JarOutputStream
 
 class ClassGroup private constructor(nodes: MutableList<ClassNode>) : MutableList<ClassNode> by nodes {
 
-    constructor() : this(mutableListOf())
-
-    operator fun get(name: String): ClassNode? = this.firstOrNull { it.name == name }
-
-    private fun init() {
-        this.forEach { c ->
-            c.setGroup(this)
-            c.init()
-        }
-    }
-
-    fun rebuild() {
-        rebuildClasses()
-        rebuildMethods()
-        rebuildFields()
-
+    init {
         this.init()
     }
 
+    /**
+     * Initializes the class group.
+     */
+    fun init() {
+        this.forEach { it.init(this) }
+    }
+
+    operator fun get(name: String): ClassNode? = this.firstOrNull { it.name == name }
+
     fun toJar(file: File) {
-        if(file.exists()) {
-            file.delete()
-        }
+        if(file.exists()) file.delete()
 
         val jos = JarOutputStream(FileOutputStream(file))
+
         this.forEach {
             jos.putNextEntry(JarEntry(it.name + ".class"))
-            val writer = ClassWriter(ClassWriter.COMPUTE_MAXS)
+
+            val writer = ClassWriter(0)
             it.accept(writer)
+
             jos.write(writer.toByteArray())
             jos.closeEntry()
         }
+
         jos.close()
     }
 
@@ -50,21 +46,20 @@ class ClassGroup private constructor(nodes: MutableList<ClassNode>) : MutableLis
 
         fun fromJar(file: File): ClassGroup {
             val nodes = mutableListOf<ClassNode>()
+
             JarFile(file).use { jar ->
                 jar.entries().asSequence()
                     .filter { it.name.endsWith(".class") }
                     .forEach {
                         val node = ClassNode()
-                        val reader = ClassReader(jar.getInputStream(it))
-                        reader.accept(node, ClassReader.SKIP_FRAMES)
+                        val reader = ClassReader(jar.getInputStream(JarEntry(it)))
+                        reader.accept(node, 0)
+
                         nodes.add(node)
                     }
             }
 
-            val group = ClassGroup(nodes)
-            group.init()
-
-            return group
+            return ClassGroup(nodes)
         }
     }
 }
